[base] Assume no Sophos URLs means use the default, and no System SSL path means use SUL's built-in system cert finder.
Tidy int -> size_t for sizes.
Change unit test to match new behaviour.

+++ base/modules/SulDownloader/WarehouseRepository.cpp
@@ -184,12 +184,12 @@ namespace SulDownloader
 
         SelectedResultsIndexes selectedIndexes = selection.selectProducts(productMetadataList);
 
-        for ( int index : selectedIndexes.selected)
+        for ( size_t index : selectedIndexes.selected)
         {
             LOGSUPPORT("Product will be downloaded: " << productMetadataList[index].getLine());
         }
 
-        for (  int index : selectedIndexes.notselected)
+        for (  size_t index : selectedIndexes.notselected)
         {
             auto & productPair = productInformationList[index];
             if(!SULUtils::isSuccess(SU_removeProduct(productPair.first)))
